⚡️ convert author profile img to webp
